Follow duck pattern for root component
