+++ generators/app/templates/client/src/containers/root/root.container.js
@@ -1,0 +1,15 @@
+import Root from './root.component';
+import { connect } from 'react-redux';
+
+function mapStateToProps(state) {
+  return {};
+}
+
+function mapDispatchToProps(dispatch) {
+  return {};
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(Root);
